+++ loopd/swapclient_server.go
@@ -21,6 +21,8 @@ import (
 	"github.com/lightningnetwork/lnd/lnwire"
 	"github.com/lightningnetwork/lnd/queue"
 	"github.com/lightningnetwork/lnd/routing/route"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 const (
@@ -702,14 +704,23 @@ func rpcToRule(rule *looprpc.LiquidityRule) (*liquidity.ThresholdRule, error) {
 func (s *swapClientServer) SuggestSwaps(ctx context.Context,
 	_ *looprpc.SuggestSwapsRequest) (*looprpc.SuggestSwapsResponse, error) {
 
-	swaps, err := s.liquidityMgr.SuggestSwaps(ctx, false)
-	if err != nil {
+	suggestions, err := s.liquidityMgr.SuggestSwaps(ctx, false)
+	switch err {
+	case liquidity.ErrNoRules:
+		return nil, status.Error(codes.FailedPrecondition, err.Error())
+
+	case nil:
+
+	default:
 		return nil, err
 	}
 
-	var loopOut []*looprpc.LoopOutRequest
+	var (
+		loopOut      []*looprpc.LoopOutRequest
+		disqualified []*looprpc.Disqualified
+	)
 
-	for _, swap := range swaps {
+	for _, swap := range suggestions.OutSwaps {
 		loopOut = append(loopOut, &looprpc.LoopOutRequest{
 			Amt:                 int64(swap.Amount),
 			OutgoingChanSet:     swap.OutgoingChanSet,
@@ -722,9 +733,69 @@ func (s *swapClientServer) SuggestSwaps(ctx context.Context,
 		})
 	}
 
+	for id, reason := range suggestions.DisqualifiedChans {
+		autoloopReason, err := rpcAutoloopReason(reason)
+		if err != nil {
+			return nil, err
+		}
+
+		exclChan := &looprpc.Disqualified{
+			Reason:    autoloopReason,
+			ChannelId: id.ToUint64(),
+		}
+		disqualified = append(disqualified, exclChan)
+	}
+
 	return &looprpc.SuggestSwapsResponse{
-		LoopOut: loopOut,
+		LoopOut:      loopOut,
+		Disqualified: disqualified,
 	}, nil
+}
+
+func rpcAutoloopReason(reason liquidity.Reason) (looprpc.AutoReason, error) {
+	switch reason {
+	case liquidity.ReasonNone:
+		return looprpc.AutoReason_AUTO_REASON_UNKNOWN, nil
+
+	case liquidity.ReasonBudgetNotStarted:
+		return looprpc.AutoReason_AUTO_REASON_BUDGET_NOT_STARTED, nil
+
+	case liquidity.ReasonSweepFees:
+		return looprpc.AutoReason_AUTO_REASON_SWEEP_FEES, nil
+
+	case liquidity.ReasonBudgetElapsed:
+		return looprpc.AutoReason_AUTO_REASON_BUDGET_ELAPSED, nil
+
+	case liquidity.ReasonInFlight:
+		return looprpc.AutoReason_AUTO_REASON_IN_FLIGHT, nil
+
+	case liquidity.ReasonSwapFee:
+		return looprpc.AutoReason_AUTO_REASON_SWAP_FEE, nil
+
+	case liquidity.ReasonMinerFee:
+		return looprpc.AutoReason_AUTO_REASON_MINER_FEE, nil
+
+	case liquidity.ReasonPrepay:
+		return looprpc.AutoReason_AUTO_REASON_PREPAY, nil
+
+	case liquidity.ReasonFailureBackoff:
+		return looprpc.AutoReason_AUTO_REASON_FAILURE_BACKOFF, nil
+
+	case liquidity.ReasonLoopOut:
+		return looprpc.AutoReason_AUTO_REASON_LOOP_OUT, nil
+
+	case liquidity.ReasonLoopIn:
+		return looprpc.AutoReason_AUTO_REASON_LOOP_IN, nil
+
+	case liquidity.ReasonLiquidityOk:
+		return looprpc.AutoReason_AUTO_REASON_LIQUIDITY_OK, nil
+
+	case liquidity.ReasonBudgetInsufficient:
+		return looprpc.AutoReason_AUTO_REASON_BUDGET_INSUFFICIENT, nil
+
+	default:
+		return 0, fmt.Errorf("unknown autoloop reason: %v", reason)
+	}
 }
 
 // processStatusUpdates reads updates on the status channel and processes them.