+++ release_notes.md
@@ -24,6 +24,14 @@ This file tracks release notes for the loop client.
   baked one with the exact permissions needed for Loop. If the now deprecated
   flag/option `--lnd.macaroondir` is used, it will fall back to use only the
   `admin.macaroon` from that directory.
+* The rules used for autoloop have been relaxed to allow autoloop to dispatch
+  swaps even if there are manually initiated swaps that are not limited to a
+  single channel in progress. This change was made to allow autoloop to coexist
+  with manual swaps. 
+* The `SuggestSwaps` endpoint has been updated to include reasons that indicate
+  why the Autolooper is not currently dispatching swaps for the set of rules
+  that the client is configured with. See the [autoloop documentation](docs/autoloop.md) for a
+  detailed explanations of these reasons.
 
 #### Breaking Changes
 * The `AutoOut`, `AutoOutBudgetSat` and `AutoOutBudgetStartSec` fields in the
@@ -31,5 +39,8 @@ This file tracks release notes for the loop client.
   been renamed to `Autoloop`, `AutoloopBudgetSat` and `AutoloopBudgetStartSec`. 
 * The `autoout` flag for enabling automatic dispatch of loop out swaps has been
   renamed to `autoloop` so that it can cover loop out and loop in.
+* The `SuggestSwaps` rpc call will now fail with a `FailedPrecondition` grpc
+  error code if no rules are configured for the autolooper. Previously the rpc
+  would return an empty response.
 
 #### Bug Fixes